changed x and y coordinate to double precision
git-svn-id: https://svn.smr.arpa.emr.it/vol7d/trunk@998 500a5620-302c-0410-bb8c-d54761bbe776

+++ examples/example_space_utilities.F90
@@ -13,7 +13,7 @@ integer,dimension(:),allocatable :: seed
 integer :: k,i
 
 integer,parameter :: ndp=10000
-real, DIMENSION(ndp) :: x,y 
+double precision, DIMENSION(ndp) :: x,y 
 integer ::  status
 character(len=512):: a_name
 INTEGER :: category, ier
